# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portofolio Naufal - Product API",
        "description": "Product API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://product-app.test/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/attachment": {
            "get": {
                "tags": [
                    "Attachment"
                ],
                "summary": "List of Attachment",
                "operationId": "a14619aa3df1f832abf9cffa2d56c571",
                "parameters": [
                    {
                        "name": "filter[original_name]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "original_name"
                    },
                    {
                        "name": "filter[keyword]",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "keyword"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "original_name"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    },
                    {
                        "name": "rows",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Store Attachment",
                "operationId": "2a340a9c8c2fa0dc0cb1da77014e7935",
                "requestBody": {
                    "description": "Body",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "service_name": {
                                        "type": "string"
                                    },
                                    "file[]": {
                                        "description": "field file ini bertype array of file upload",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data sukses disimpan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "service_name": {
                                            "type": "array",
                                            "items": {
                                                "example": [
                                                    "service_name field is required."
                                                ]
                                            }
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "example": {
                                                    "file": "the file field is required"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/attachment/{attachment}": {
            "get": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Attachment details",
                "operationId": "686d610268cd85e1fb0f24b28f273b09",
                "parameters": [
                    {
                        "name": "attachment",
                        "in": "path",
                        "description": "Attachment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Update Attachment",
                "operationId": "8300c633752c671b3424e1fedf2befa0",
                "parameters": [
                    {
                        "name": "attachment",
                        "in": "path",
                        "description": "Attachment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Body",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data sukses disimpan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "file": {
                                            "type": "array",
                                            "items": {
                                                "example": [
                                                    "file_name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Attachment Removal",
                "operationId": "b6882f55b1e3af94143f9dfb1aa999c4",
                "parameters": [
                    {
                        "name": "attachment",
                        "in": "path",
                        "description": "Attachment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data sukses dihapus."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/attachment/{attachment}/download": {
            "get": {
                "tags": [
                    "Attachment"
                ],
                "summary": "Download Attachment",
                "operationId": "311cef4ed6238b17297ee7d1e9d2adec",
                "parameters": [
                    {
                        "name": "attachment",
                        "in": "path",
                        "description": "Attachment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "404": {
                        "description": "error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "File tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}